ai-interviewer-ace/backend/requirements.txt
# Backend dependencies for HireGage AI Interviewer
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
openai==1.3.5
python-multipart==0.0.6
httpx==0.24.1
asyncio==3.4.3
supabase==1.0.4
sqlalchemy==2.0.23
asyncpg==0.28.0
pytest==7.4.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
alembic==1.12.1
pytest==7.4.3